Fix issue in string conversion
peekCStringLen uses the local encoding to convert the bytes to an
Haskell String. This can result in bytes dropped by the conversion if
the sequence of bytes is not interpretable in the current encoding.

We should actually use ByteString here as the value is not necessarily a
String.

Using the version peekCAStringLen ensures all the actual bytes end up
in the Haskell String. Then one can convert it to a ByteString with
Data.ByteString.Char8